+++ src/main/java/org/fsg1/fmms/backend/app/Main.java
@@ -1,4 +1,4 @@
-package resources;
+package org.fsg1.fmms.backend.resources;
 
 import org.glassfish.grizzly.http.server.HttpServer;
 import org.glassfish.jersey.grizzly2.httpserver.GrizzlyHttpServerFactory;
@@ -9,11 +9,15 @@ import java.net.URI;
 
 /**
  * Main class.
- *
  */
-public class Main {
-    // Base URI the Grizzly HTTP server will listen on
-    static final String BASE_URI = "http://localhost:8080/fmms/";
+public final class Main {
+
+    /**
+     * Private constructor.
+     * Class should never be instantiated.
+     */
+    private Main() {
+    }
 
     /**
      * Starts Grizzly HTTP server exposing JAX-RS resources defined in this application.
@@ -39,14 +43,13 @@ public class Main {
 
     /**
      * Main method.
-     * @param args
-     * @throws IOException
+     * @param args Command line arguments
+     * @throws IOException if an error occurs while attempting to start the server.
      */
-    public static void main(String[] args) throws IOException {
+    public static void main(final String[] args) throws IOException {
         System.out.println("Booting server ...");
 
         final HttpServer server = startServer();
-
 
         // register shutdown hook
         Runtime.getRuntime().addShutdownHook(new Thread(() -> {
