--- src/main/resources/config/database.properties
@@ -1,3 +1,0 @@
-db_user=fmms
-db_passwd=
-db=localhost:5432/fmms
