--- src/main/resources/config/server.properties
@@ -1,3 +1,0 @@
-host=0.0.0.0
-port=8080
-base=/fmms
