+++ src/test/java/org/fsg1/fmms/backend/database/ConnectionTest.java
@@ -3,8 +3,10 @@ package org.fsg1.fmms.backend.database;
 import com.mockrunner.jdbc.BasicJDBCTestCaseAdapter;
 import com.mockrunner.mock.jdbc.MockParameterMap;
 import com.mockrunner.mock.jdbc.MockPreparedStatement;
+import org.fsg1.fmms.backend.resources.Configuration;
 import org.junit.Before;
 import org.junit.Test;
+import org.mockito.Mockito;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -13,14 +15,22 @@ import static org.junit.Assert.assertEquals;
 
 public class ConnectionTest extends BasicJDBCTestCaseAdapter {
 
+    Configuration configMock;
+
     @Before
     public void setupJDBC(){
         getJDBCMockObjectFactory().registerMockDriver();
+
+        configMock = Mockito.mock(Configuration.class);
+        Mockito.when(configMock.getDbString()).thenReturn("jdbc:postgresql://localhost:5432/fmms");
+        Mockito.when(configMock.getDbUser()).thenReturn("fmms");
+        Mockito.when(configMock.getDbPassword()).thenReturn("test123456");
     }
 
     @Test
     public void testExecuteQuery() throws SQLException {
-        Connection conn = new Connection();
+        Connection conn = new Connection(configMock);
+
         conn.executeQuery("param1: '?' , param2: '?'", "stringparam", 4);
         final List<MockPreparedStatement> preparedStatements = getJDBCMockObjectFactory().getMockConnection()
                 .getPreparedStatementResultSetHandler().getPreparedStatements();
@@ -34,7 +44,7 @@ public class ConnectionTest extends BasicJDBCTestCaseAdapter {
 
     @Test
     public void testSetParameters() throws SQLException {
-        Connection conn = new Connection();
+        Connection conn = new Connection(configMock);
         String query = "SELECT * FROM ?";
         Object[] params = new Object[]{"tablename", 2, 4, "fourth param"};
         conn.executeQuery(query, params);
